fcc_split/requirements.txt
xmltodict
python-dateutil